Diminui quantidade de linhas page Game
