+++ src/helpers/index.js
@@ -1,0 +1,11 @@
+export default function setInitialState() {
+  const playerObj = {
+    player: {
+      name: '',
+      assertions: '',
+      score: 0,
+      gravatarEmail: '',
+    },
+  };
+  localStorage.setItem('state', JSON.stringify(playerObj));
+}
